DOMA-1443 Resolving rebase conflicts
