+++ apps/condo/domains/acquiring/constants/integration.js
@@ -1,0 +1,5 @@
+const DEFAULT_ACQUIRING_INTEGRATION_NAME = 'default'
+
+module.exports = {
+    DEFAULT_ACQUIRING_INTEGRATION_NAME,
+}
